Create Simple Wallet
This simple wallet contract allows for depositing funds, only by sending Ether to the contract address. It includes a withdrawal function that can be called by the owner to transfer funds from the contract to a specified recipient address. The contract also provides a function to retrieve the current balance of the wallet.